requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
google-generativeai>=0.3.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.0.0
psutil>=5.9.0
flask>=2.3.0
flask-cors>=4.0.0
chromadb>=0.4.0
numpy>=1.24.0
pandas>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

